# practice-ui/docs/api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Interview Preparation Platform API",
    "description": "Complete API reference for the Interview Preparation Platform backend services. This API provides endpoints for managing interview questions, practice sessions, user analytics, and application configuration for Java, Selenium, and API testing preparation.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@interviewprep.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://seleniuminterviewprep.netlify.app/api",
      "description": "Production server"
    }
  ],
  "paths": {
    "/questions": {
      "get": {
        "tags": ["Questions Management"],
        "summary": "Get All Questions",
        "description": "Retrieve all interview questions across all categories",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of questions to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of questions to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "description": "Filter by difficulty level",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["beginner", "intermediate", "advanced"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Questions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/questions/{category}": {
      "get": {
        "tags": ["Questions Management"],
        "summary": "Get Questions by Category",
        "description": "Retrieve questions for a specific category (java, selenium, api, database, etc.)",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "description": "Question category",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["java", "selenium", "api", "database", "testng", "framework", "performance", "mobile", "security", "cicd", "leadership"]
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "description": "Filter by difficulty level",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["beginner", "intermediate", "advanced"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of questions to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category questions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryQuestionsResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique question identifier",
            "example": "q001"
          },
          "category": {
            "type": "string",
            "description": "Question category",
            "enum": ["java", "selenium", "api", "database", "testng", "framework", "performance", "mobile", "security", "cicd", "leadership"],
            "example": "java"
          },
          "question": {
            "type": "string",
            "description": "The question text",
            "example": "What is JVM?"
          },
          "answer": {
            "type": "string",
            "description": "The answer text",
            "example": "Java Virtual Machine is a runtime environment that executes Java bytecode..."
          },
          "difficulty": {
            "type": "string",
            "description": "Question difficulty level",
            "enum": ["beginner", "intermediate", "advanced"],
            "example": "beginner"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Question tags for filtering",
            "example": ["java", "basics", "jvm"]
          },
          "experienceLevel": {
            "type": "string",
            "description": "Target experience level",
            "enum": ["0-2", "3-5", "6-8", "8+"],
            "example": "0-2"
          }
        },
        "required": ["id", "category", "question", "difficulty"]
      },
      "QuestionsResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/BaseResponse"},
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Question"
                }
              },
              "total": {
                "type": "integer",
                "description": "Total number of questions available",
                "example": 155
              },
              "pagination": {
                "$ref": "#/components/schemas/Pagination"
              }
            }
          }
        ]
      },
      "CategoryQuestionsResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/BaseResponse"},
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Question"
                }
              },
              "category": {
                "type": "string",
                "example": "java"
              },
              "count": {
                "type": "integer",
                "example": 32
              }
            }
          }
        ]
      },
...
      "BaseResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request was successful"
          },
          "message": {
            "type": "string",
            "description": "Optional message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Response timestamp",
            "example": "2025-12-15T10:00:00Z"
          }
        },
        "required": ["success"]
      },
...
      "ErrorResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/BaseResponse"},
          {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "example": false
              },
              "error": {
                "type": "string",
                "description": "Error message"
              },
              "code": {
                "type": "string",
                "description": "Error code"
              },
              "details": {
                "type": "string",
                "description": "Detailed error information"
              }
            },
            "required": ["error"]
          }
        ]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer",
            "example": 100
          },
          "offset": {
            "type": "integer",
            "example": 0
          },
          "total": {
            "type": "integer",
            "example": 155
          },
          "hasMore": {
            "type": "boolean",
            "example": true
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Invalid request parameters",
              "code": "INVALID_PARAMS",
              "timestamp": "2025-12-15T10:00:00Z"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Resource not found",
              "code": "NOT_FOUND",
              "timestamp": "2025-12-15T10:00:00Z"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Internal server error",
              "code": "INTERNAL_ERROR",
              "timestamp": "2025-12-15T10:00:00Z"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId",
        "description": "Session-based authentication using cookies"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token authentication"
      }
    }
  },
  "security": [
    {
      "SessionAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
}